Bugfix: When null is set as resultsPerPage, there is a default setting for how many cards to show (which is 9). That is ignored.
Notes: Apologies. I was under the impression that null was a valid value e.g. if you don’t wanna limit cards at all. Normally values that don’t exist are represented either by undefined or the key is simply not there. Have changed the getConfig function to reflect this.